Add application specific listener
Not all information communicated by the compositor is related to some
kind of surface. One example of this is outputs, outputs is a software
representation of our physical displays.

Adding global state callbacks to the shell listener becomes awkward
because one application instance may have multiple shell's. Thus the
ergonomic way to solve this issue is to create callbacks on the
application instead.

Signed-off-by: Linus Johansen <linusen.a@gmail.com>