Add support for wl_output
The wayland compositor communciates information about the physical
displays out the clients. Aylin now listens for outputs on the wayland
socket. Using the application_listener we are able to communciate
realtime with the application instance that a new output has been
discovered.

All the outputs are also put into a linked list in aylin_application.
This gives us the ability to create a more advanced api over the
outputs.

Signed-off-by: Linus Johansen <linusen.a@gmail.com>