Add support for enriched resize information.
Previously you needed to check for resizing in the frame event. Now
aylin is capable of tell us when xdg surfaces are resized by reading out
the states sent in xdg_toplevel.configure.

This allows clients to make
more sophisticated decisions on what to do with their decorations and
alike during resizing.

Signed-off-by: Linus Johansen <linusen.a@gmail.com>